rename cli command and add more comments on the purpose and arguments
