+++ node/src/manager/commands/mod.rs
@@ -3,9 +3,9 @@ pub mod chain;
 pub mod config;
 pub mod copy;
 pub mod create;
-pub mod dump_poi;
 pub mod info;
 pub mod listen;
+pub mod poi;
 pub mod query;
 pub mod remove;
 pub mod rewind;
