+++ node/src/manager/commands/poi.rs
